V4 done. 4.1M. 1B rows in 4.0 minutes

+++ 1BillionRowChallenge/Models/AggregatedDataPointV4.cs
@@ -2,11 +2,11 @@
 
 public class AggregatedDataPointV4
 {
-    public double Min;
+    public decimal Min;
 
-    public double Max;
+    public decimal Max;
 
-    public double Sum;
+    public decimal Sum;
 
     public ushort AmountOfDataPoints;
 }
